requirements.txt
isort==5.12.0
black==23.7.0
Django==4.2.4
djangorestframework==3.14.0
django-filter==23.2
djangorestframework-simplejwt==5.2.2
